System supports deepmd/raw and deepmd/npy formats.

+++ dpdata/deepmd/comp.py
@@ -2,20 +2,26 @@ import os,glob,shutil
 import numpy as np
 from .raw import load_type
 
+def _cond_load_data(fname) :
+    tmp = None
+    if os.path.isfile(fname) :
+        tmp = np.load(fname)
+    return tmp
+
 def _load_set(folder) :
     cells = np.load(os.path.join(folder, 'box.npy'))
     coords = np.load(os.path.join(folder, 'coord.npy'))
-    eners = np.load(os.path.join(folder, 'energy.npy'))
-    forces = np.load(os.path.join(folder, 'force.npy'))
-    virs = None
-    if os.path.isfile(os.path.join(folder, 'virial.npy')) :
-        virs = np.load(os.path.join(folder, 'virial.npy'))
+    eners  = _cond_load_data(os.path.join(folder, 'energy.npy'))
+    forces = _cond_load_data(os.path.join(folder, 'force.npy'))
+    virs   = _cond_load_data(os.path.join(folder, 'virial.npy'))
     return cells, coords, eners, forces, virs
 
-def to_system_data(folder, type_map = None) :
+def to_system_data(folder, 
+                   type_map = None, 
+                   labels = True) :
+    # data is empty
     data = load_type(folder, type_map = type_map)
     data['orig'] = np.zeros([3])
-    data['virials'] = []
     sets = glob.glob(os.path.join(folder, 'set.*'))
     all_cells = []
     all_coords = []
@@ -27,14 +33,19 @@ def to_system_data(folder, type_map = None) :
         nframes = np.reshape(cells, [-1,3,3]).shape[0]
         all_cells.append(np.reshape(cells, [nframes,3,3]))
         all_coords.append(np.reshape(coords, [nframes,-1,3]))
-        all_eners.append(np.reshape(eners, [nframes]))
-        all_forces.append(np.reshape(forces, [nframes,-1,3]))
-        if virs is not None and len(virs) > 0:
-            virs = all_virs.append(np.reshape(virs, [nframes,3,3]))
+        if labels:
+            if eners is not None and len(eners) > 0:
+                all_eners.append(np.reshape(eners, [nframes]))
+            if forces is not None and len(forces) > 0:
+                all_forces.append(np.reshape(forces, [nframes,-1,3]))
+            if virs is not None and len(virs) > 0:
+                all_virs.append(np.reshape(virs, [nframes,3,3]))
     data['cells'] = np.concatenate(all_cells, axis = 0)
-    data['coords'] = np.concatenate(all_coords, axis = 0)
-    data['energies'] = np.concatenate(all_eners, axis = 0)
-    data['forces'] = np.concatenate(all_forces, axis = 0)
+    data['coords'] = np.concatenate(all_coords, axis = 0)    
+    if len(all_eners) > 0 :
+        data['energies'] = np.concatenate(all_eners, axis = 0)
+    if len(all_forces) > 0 :
+        data['forces'] = np.concatenate(all_forces, axis = 0)
     if len(all_virs) > 0:
         data['virials'] = np.concatenate(all_virs, axis = 0)
     return data
@@ -59,12 +70,15 @@ def dump(folder,
     nframes = data['cells'].shape[0]
     cells  = np.reshape(data['cells'],    [nframes,  9]).astype(comp_prec)
     coords = np.reshape(data['coords'],   [nframes, -1]).astype(comp_prec)
-    eners  = np.reshape(data['energies'], [nframes    ]).astype(comp_prec)
-    forces = np.reshape(data['forces'],   [nframes, -1]).astype(comp_prec)
-    if len(data['virials']) > 0 :
-        virials = np.reshape(data['virials'],   [nframes, 9]).astype(comp_prec)
-    else :
-        virials = []
+    eners = None
+    forces = None
+    virials = None
+    if 'energies' in data:
+        eners   = np.reshape(data['energies'], [nframes    ]).astype(comp_prec)    
+    if 'forces' in data:
+        forces  = np.reshape(data['forces'],   [nframes, -1]).astype(comp_prec)
+    if 'virials' in data :
+        virials = np.reshape(data['virials'],  [nframes,  9]).astype(comp_prec)
     if 'atom_pref' in data:
         atom_pref = np.reshape(data['atom_pref'], [nframes, -1]).astype(comp_prec)
     # dump frame properties: cell, coord, energy, force and virial
@@ -78,9 +92,11 @@ def dump(folder,
         os.makedirs(set_folder)
         np.save(os.path.join(set_folder, 'box'),        cells  [set_stt:set_end])
         np.save(os.path.join(set_folder, 'coord'),      coords [set_stt:set_end])
-        np.save(os.path.join(set_folder, 'energy'),     eners  [set_stt:set_end])
-        np.save(os.path.join(set_folder, 'force'),      forces [set_stt:set_end])
-        if len(virials) > 0:
+        if eners is not None:
+            np.save(os.path.join(set_folder, 'energy'), eners  [set_stt:set_end])
+        if forces is not None:
+            np.save(os.path.join(set_folder, 'force'),  forces [set_stt:set_end])
+        if virials is not None:
             np.save(os.path.join(set_folder, 'virial'), virials[set_stt:set_end])
         if 'atom_pref' in data:
             np.save(os.path.join(set_folder, "atom_pref"), atom_pref[set_stt:set_end])
